// frontend/src/domains/shared/phone/generated/country-data.ts
export interface Country {
  variant: string;
  alpha2: string;
  alpha3: string;
  dialCode: number;
  name: string;
  flag: string;
}